Add back support for Go 1.14
Adds simple reimplementation of fillBytesInt for go 1.14 support

Fix #50
